tkinter/requirements.txt
Flask==2.1.1
PyPDF2==1.26.0
Pillow==9.0.0